Model features as a dict
Seems more Pythonic, bakes some assumptions in though.
For example, we kind of gloss over what happens if we see more than
one instance of a feature in the glozz reader.  Something would
get clobbered.

Ideally we would make the reader catch this and percolate up some
kind of warning or exception?