Introduce flat graph representation
The flat representation does not rely on the notion of a
hypergraph and instead represents CDUs as nodes that
point to their contents. I'm not sure why exactly I think
this will be useful, but it's worth exploring.

In the interest of sharing useful functions between the two
graph representations, I've split things apart a bit so that we have

  AttrsMixin - common node/edge attributes like 'is_cdu'
  |
  +-- HypergraphBase -+-- Graph
  |                   |
  |                   +-- FakeGraph (for testing)
  |
  +-- (DigraphBase)--+--- FlatGraph

DigraphBase doesn't actually exist but it probably should