educe.external tree for dealing w 3rd party tools
Not sure if we really want this in educe or not.
In most basic form this is really about reading in annotations
produced by other tools and providing a way to refer them
from other annotations.

Here we also throw in some facilities for running the POS
tagger and reading back the results