include orphan EDUs in graph
The graph from doc criteria are a bit subtle. We want to include
some non-EDU items as nodes because this can be used to highlight
annotation errors (eg. pointing to Turns); but at the same time,
we don't necessarily want to include every unit annotation in sight
so to handle this, we previously only included unit annotations that
had some sort of pointer to them (on the discourse level).

This turns out to be too strict, because it excludes otherwise
legitimate EDUs that happen to be orphans.