ENH exclude stage from STAC global id (nb breaks compat)
This is code-compatible but the behaviour may be incompatible
with prior uses of the library.

This subtly changes the notion of id so that it's not unique
across documents, but then it never *really* was anyway since
we were excluding the annotators.

The new notion of global id allows us to reflect some smoothly
a notion of a 'moral' identifier that an object can have the
same id across different documents.

Bit more background
-------------------
Recall also that the main problem global ids are supposed to solve
is that local ids cannot really be trusted to be unique (they may
be automatically generated and if done badly may repeat across
totally different documents).  So we need identifiers that are unique
enough that different docs/subdocs get different identifiers.

To recap:

- local : not unique enough
- doc-subdoc-local:
- doc-subdoc-stage-local: (prior to this patch) too unique
- doc-subdoc-stage-annotator-local: way too unique

If in the future it becomes really really important to be able
to talk about identifiers and to treat the same moral object as
being different across different layers, we can introduce the
doc-subdoc-stage-annotator ids too.